follow機能を切り出すためのmoduleの導入
