+++ app/services/comfirm_following_service.rb
@@ -1,0 +1,5 @@
+module ComfirmFollowingService
+  def following?(user)
+    following_relationships.find_by(following_id: user.id)
+  end
+end
