+++ app/services/follow_user_service.rb
@@ -1,0 +1,5 @@
+module FollowUserService
+  def follow(user)
+    following_relationships.create!(following_id: user.id)
+  end
+end
